refactor: styling
