--- .husky/pre-push
@@ -1,2 +1,0 @@
-npm run test
-npx vitest run --coverage
